Change to query on the most recent date --max(date)
1) slLib\colWidth()
Change font weight to adjust cell width for font sizes Pt17 and Pt15
--elif font_size > 16: weight = 0.75
--elif font_size > 14: weight = 0.80

2) slDeck\loadDbTables() - Remove 'PRIMARY KEY' setting
from sw_name on 'SwitchSummary' and 'SwitchPortUsage'
to allow loading multiple reports with the same fabrics but different
dates.
The setting was forcing the swhitch name to be unique and not loading all
records
from the csv files.

3) slDeck_multi and single:
Add to slides (zonning, switches and errors), an SQL WHERE statement,
  'and date = (SELECT max(date) FROM <table>)'
  to grab only the most recent report data for non-compared decks (single
  and combined)

  Add to slide (Fabric)
  ON s.sw_name = ports.sw_name and s.date = ports.date
  WHERE s.date = (SELECT max(date) FROM switches)